# design/publisher-openapi.yaml
openapi: 3.1.0
info:
  title: TEA Publisher API
  version: 0.11.0-draft
  description: |
    **Status: draft v0.11, for discussion. opentea's own server-side implementation
    exists (`internal/publisher`) but this document is still a draft, not approved by
    CycloneDX or any TEA working group.**

    v0.11: TEA 1.0 conformance — new `uploadArtifactSignatureFile` operation, alongside a
    fix to `formats[].url`/`formats[].signatureUrl` semantics on the consumer API side
    (`spec/openapi.yaml`, not this document): both are now populated only for genuinely
    external locations, never opentea's own storage; self-hosted content and signatures
    are instead retrieved via four new consumer-API download endpoints
    (`/artifact/{uuid}/{latest|version}/download` and their `/signature/download`
    counterparts — strong `ETag`/`If-None-Match`, `HEAD`, `Content-Location`, `mediaType`/
    `Accept`-based format selection, `302` redirect for external locations, `406` for no
    matching format, `404` for a concealed/nonexistent artifact and, separately, for a real
    format with no signature published). This is a behavior change for any existing
    self-hosted-content consumer relying on `formats[].url` being populated. `TODO.md`
    records the deferred question of whether `release_distribution.url` (which has no
    TEA-hosted download endpoint at all, unlike artifact formats) needs the same treatment.

...
    here — see `TODO.md`'s **Publisher API** entry.

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /productReleases/{productReleaseUuid}/components:
    post:
      operationId: linkComponent
      tags: [Components]
      summary: Link a component to a product release, optionally pinned to a specific component release.
      parameters:
        - $ref: '#/components/parameters/ProductReleaseUuid'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/component-ref' }
      responses:
        '204': { description: Linked. }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- Artifacts
  /artifacts:
    post:
      operationId: createArtifact
      tags: [Artifacts]
      summary: >-
        Create an artifact (metadata only — file content and evidence are attached by the
        two operations below). Corresponds to oej's workflow phase 1/2 boundary: this
        operation itself is not a signing event.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/artifact-create' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/artifact' }
  /artifacts/{artifactUuid}/{version}/files:
    post:
      operationId: uploadArtifactFile
      tags: [Artifacts]
      summary: >-
        Upload the artifact's actual file content for one of its formats. Server computes
        checksums from the uploaded bytes (see info.description, deviation 4) — a caller
        never supplies them directly. Rejected with 409 once evidence has been submitted
        for this artifact version (`submitArtifactEvidence`) — file content is frozen once
        validated, a subsequent correction belongs under a new artifact version, not a
        file swap on an already-signed one (v0.10, found by external security review:
        docs/security-review-publisher-design-260828.md finding 1 — a re-upload after
        evidence submission would otherwise silently invalidate the signature's meaning
        without the stored evidence bundle reflecting it). Subject to a per-upload size
        limit and a per-principal outstanding-bytes/count quota (design/publisher-service.md
        §15, §15.1) — neither has a chosen default here. A well-behaved caller (the
        reference client, §14.1) should check its own file size against a sane limit
        before attempting this call at all, but that is a courtesy, not a substitute for
        this operation's own enforcement (§15.1).
      parameters:
        - $ref: '#/components/parameters/ArtifactUuid'
        - $ref: '#/components/parameters/ArtifactVersion'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [mediaType, file]
              properties:
                mediaType:
                  type: string
                  description: >-
                    Selects which of the artifact's formats this file belongs to, by the
                    same `mediaType` value that format was created with
                    (`artifact-format-create.mediaType`) — not a positional array index
                    (v0.10, found by external security review:
                    docs/security-review-publisher-design-260828.md finding 14 — array
                    order is not a durable identifier). `404` if no format matches; `400`
                    if more than one format shares this `mediaType` (ambiguous — this
                    draft does not yet define a server-assigned per-format id to
                    disambiguate that case, a narrower, still-open version of the same
                    finding).
                file: { type: string, format: binary }
      responses:
        '204': { description: Uploaded. }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Evidence has already been submitted for this artifact version.
        '413':
          description: File exceeds the target's per-upload size limit (§15.1).
        '429':
          description: >-
            This principal's *unused*-artifact quota (bytes and/or count, not referenced
            by any collection) is exhausted (§15.1) — distinct from a used-storage
            capacity ceiling, which is an operational/provisioning concern, not a
            per-request rejection at all. One of three plausible status codes for this
            specific case (429, 507, 403 all have an argument, §15.1's own open
            question); not settled here.
  /artifacts/{artifactUuid}/{version}/signature/files:
    post:
      operationId: uploadArtifactSignatureFile
      tags: [Artifacts]
      summary: >-
        Upload a detached signature for one of the artifact's formats, selected by
        `mediaType` the same way as `uploadArtifactFile` (v0.11, TEA 1.0 conformance —
        `spec/openapi.yaml`'s `formats[].signatureUrl` is now documented as always an
        external location; a target may instead host the signature itself, retrievable via
        the TEA consumer API's own `.../signature/download` operation). Unlike
        `uploadArtifactFile`, never rejected once evidence has been submitted for this
        artifact version — evidence attests to the format's content checksum, which a
        signature upload never changes. Subject to the same per-upload size limit and
        per-principal quota as `uploadArtifactFile` (§15, §15.1).
      parameters:
        - $ref: '#/components/parameters/ArtifactUuid'
        - $ref: '#/components/parameters/ArtifactVersion'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [mediaType, file]
              properties:
                mediaType:
                  type: string
                  description: >-
                    Selects which of the artifact's formats this signature belongs to, by
                    the same `mediaType` value that format was created with — identical
                    selection rule and `404`/`400`-ambiguous behavior as
                    `uploadArtifactFile`.
                file: { type: string, format: binary }
      responses:
        '204': { description: Uploaded. }
        '404': { $ref: '#/components/responses/NotFound' }
        '413':
          description: File exceeds the target's per-upload size limit (§15.1).
        '429':
          description: This principal's outstanding-bytes/count quota is exhausted (§15.1).
  /artifacts/{artifactUuid}/{version}/evidence/prepare:
    post:
      operationId: prepareArtifactEvidence
      tags: [Artifacts]
      summary: >-
        Prepare-to-sign (new in this revision — see design/publisher-service.md §9.2 for
        why artifacts need this, not just collections). The target server is the sole
        authority on the artifact's exact RFC 8785 canonical form (it's what verification
        will recompute against regardless); this returns that form's digest so the caller
        never has to reimplement canonicalization independently. No key material involved
        in either direction. Callable repeatedly — reflects current state each time, does
        not lock anything; only the submit operation below finalizes anything.
      parameters:
        - $ref: '#/components/parameters/ArtifactUuid'
        - $ref: '#/components/parameters/ArtifactVersion'
      responses:
        '200':
          description: The current artifact and its digest to sign.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/prepare-artifact-evidence-response' }
        '404': { $ref: '#/components/responses/NotFound' }
  /artifacts/{artifactUuid}/{version}/evidence:
    post:
      operationId: submitArtifactEvidence
      tags: [Artifacts]
      summary: >-
        Artifact Validation (oej workflow phase 3) — submit the evidence package (§9.3)
        produced by signing prepareArtifactEvidence's digest. The typical caller is a
        CI/CD system, not a human through a GUI — this operation is deliberately usable
        headlessly, with no dependency on any collection-drafting workflow. The server
        re-derives the digest itself from the artifact's actual current content (§9.1 step
        3 — never trusts a caller-supplied digest, and this may legitimately differ from
        what prepare returned if the artifact changed in between) and rejects (does not
        store) anything that fails verification — mirrors opentea's existing
        verify-before-store contract (`internal/admin/evidencebundle.go`). No
        consumer-spec equivalent (see info.description) — produces an `evidence-bundle`,
        reusing the trust-architecture extension already present on `artifact-format`.
      parameters:
        - $ref: '#/components/parameters/ArtifactUuid'
        - $ref: '#/components/parameters/ArtifactVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/evidence-submission' }
      responses:
        '201':
          description: Verified and stored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/evidence-bundle' }
        '400':
          description: Verification failed (bad signature, digest mismatch, or expired/malformed certificate) — nothing was stored.
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- Lifecycle (CLE)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Matches spec/publisher's old precedent (bearer/basic) and opentea's own
        `/admin/v1` token model. Basic auth intentionally not offered here — bearer only,
        for a write API this sensitive.
  parameters:
...
    ProductReleaseUuid:
      name: productReleaseUuid
      in: path
      required: true
      schema: { $ref: '#/components/schemas/uuid' }
...
    ArtifactUuid:
      name: artifactUuid
      in: path
      required: true
      schema: { $ref: '#/components/schemas/uuid' }
    ArtifactVersion:
      name: version
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
...
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/error-response' }
  schemas:
...
    uuid:
      type: string
      format: uuid
    date-time:
      type: string
      format: date-time
...
    component-ref:
      type: object
      description: A reference to a TEA component or specific component release
      properties:
        uuid:
          description: A unique identifier for the TEA component
          $ref: '#/components/schemas/uuid'
        release:
          description: >-
            Optional UUID of a specific release included in the product in the case where
            the product always include a specific release of a component.
          $ref: '#/components/schemas/uuid'
      required: [uuid]
...
    artifact:
      type: object
      description: A security-related document
      properties:
        uuid:
          description: Together with version, uniquely identifies the TEA Artifact.
          $ref: '#/components/schemas/uuid'
        version:
          type: integer
          default: 1
          description: Together with uuid, uniquely identifies the TEA Artifact.
        name:
          type: string
          description: Name of TEA Artifact
        type:
          description: Type of TEA Artifact
          $ref: '#/components/schemas/artifact-type'
        createdDate:
          description: The date when the TEA Artifact revision was created.
          $ref: '#/components/schemas/date-time'
        distributionIds:
          type: array
          description: >-
            List of TEA Component Release distributions this Artifact applies to. If
            absent or empty, applies to all distributions.
          items: { $ref: '#/components/schemas/uuid' }
        formats:
          type: array
          description: List of objects with the same content, but in different formats.
          items: { $ref: '#/components/schemas/artifact-format' }
      required: [uuid, type, formats]
...
    error-response:
      type: object
      description: Error response
      additionalProperties: false
      properties:
        error: { type: string }
      required: [error]

    # ================================================================
    # Trust-architecture extension (already-adopted precedent, not new — see
    # artifact-format's x-deviation above). Field shapes mirror pkg/tea/trust.go, which
    # mirrors oej/tea-trust-architecture's evidence-bundle-schema.json.
    # ================================================================
...
    artifact-create:
      type: object
      description: Same as `artifact`, minus the server-assigned uuid/version/createdDate.
      x-deviation: 'Omits uuid, version (server-assigned, defaults to 1), createdDate. See info.description, deviation 2.'
      properties:
        name: { type: string }
        type: { $ref: '#/components/schemas/artifact-type' }
        distributionIds:
          type: array
          items: { $ref: '#/components/schemas/uuid' }
        formats:
          type: array
          items: { $ref: '#/components/schemas/artifact-format-create' }
      required: [type, formats]
...
    evidence-submission:
      type: object
      description: >-
        An **evidence package** (design/publisher-service.md §9.3 — same concept as
        `evidence-bundle`, request-shaped): everything needed to verify one signature over
        one object. The single request shape used everywhere a signature is submitted in
        this API — `commitCollectionDraft` takes this directly, with nothing added,
        since approval (§7.9) moved to its own tracked decision.
      required: [objectDigestValue, signatureFormat, signatureValue, certificatePem]
      properties:
        objectDigestValue:
          type: string
          description: >-
            Hex SHA-256 of RFC 8785-canonicalized object content, as returned by the
            matching prepare operation. The server independently re-derives this from the
            object's actual current state and rejects a mismatch (§9.1 step 3) — this
            field is the caller's own record of what it believed it was signing, checked
            against reality, never trusted on its own.
        objectMediaType: { type: string }
        objectLocation: { type: string, format: uri }
        signatureFormat:
          type: string
          description: >-
            Which format depends on the signing mode (design/publisher-service.md §9.5):
            `jws-detached` fits an ephemeral, self-signed software key (the only format
            opentea's Phase 1 actually verifies today) — the lightweight, same-session
            case. `cms-detached` is the natural fit for a Web PKI certificate,
            HSM/PKCS#11-backed or air-gapped signing, since that's what most such tooling
            already produces natively — reserved in `internal/trust.SignatureFormat`'s
            vocabulary, not yet implemented anywhere. `dsse-envelope`/`cose-sign1` remain
            reserved, no concrete use case identified yet.
        signatureValue:
          type: string
          description: >-
            Base64-encoded signature over the digest bytes, in signatureFormat's encoding.
            Algorithm follows from the signing certificate's own key type — Ed25519 for
            opentea's Phase 1 ephemeral certificates, whatever the Web PKI certificate
            actually uses (commonly RSA or ECDSA) for §9.5's Mode 2 — not fixed by this
            schema.
        certificatePem:
          type: string
          description: The leaf (signing) certificate.
        certificateChain:
          type: array
          items: { type: string }
          description: >-
            Intermediate CA certificates, leaf-to-root order, needed to validate
            certificatePem against a trust store (§9.5's Mode 2) — normally empty/absent
            for opentea Phase 1's self-signed certificates, which chain to nothing.
    prepare-artifact-evidence-response:
      type: object
      description: Symmetric with prepare-commit-response, for artifacts — see prepareArtifactEvidence.
      required: [digestToSign, artifact]
      properties:
        digestToSign:
          type: string
          description: Hex SHA-256 of RFC 8785-canonicalized artifact — sign exactly these bytes.
        artifact: { $ref: '#/components/schemas/artifact' }
...